Moved some things around made the webhook controller a base class for all other webhook controllers
